Try to fix Access Denied while shelling of Chem4WordUpdater.exe

+++ src/Chem4Word.V3/UI/AutomaticUpdate.cs
@@ -43,18 +43,6 @@ namespace Chem4Word.UI
 
         public System.Windows.Point TopLeft { get; set; }
 
-        //private const int CP_NOCLOSE_BUTTON = 0x200;
-
-        //protected override CreateParams CreateParams
-        //{
-        //    get
-        //    {
-        //        CreateParams myCp = base.CreateParams;
-        //        myCp.ClassStyle = myCp.ClassStyle | CP_NOCLOSE_BUTTON;
-        //        return myCp;
-        //    }
-        //}
-
         public AutomaticUpdate(IChem4WordTelemetry telemetry)
         {
             string module = $"{_product}.{_class}.{MethodBase.GetCurrentMethod().Name}()";
@@ -127,18 +115,37 @@ namespace Chem4Word.UI
             try
             {
                 string source = Path.Combine(Globals.Chem4WordV3.AddInInfo.DeploymentPath, "Chem4WordUpdater.exe");
-                string destination = Path.Combine(Path.GetTempPath(), "Chem4WordUpdater.exe");
-                File.Copy(source, destination, true);
+                string userTempPath = Path.GetTempPath();
+                if (Directory.Exists(userTempPath))
+                {
+                    string destination = Path.Combine(userTempPath, "Chem4WordUpdater.exe");
+                    File.Copy(source, destination, true);
 
-                _telemetry.Write(module, "AutomaticUpdate", "Starting download of " + _downloadUrl);
+                    if (File.Exists(destination)
+                        && new FileInfo(destination).Length > 0)
+                    {
+                        _telemetry.Write(module, "AutomaticUpdate", "Starting updater, to download " + _downloadUrl);
 
-                ProcessStartInfo psi = new ProcessStartInfo();
-                psi.Arguments = _downloadUrl;
-                psi.FileName = destination;
-                Process.Start(psi);
+                        ProcessStartInfo psi = new ProcessStartInfo();
+                        psi.WorkingDirectory = userTempPath;
+                        psi.Arguments = _downloadUrl;
+                        psi.FileName = destination;
+                        Process.Start(psi);
 
-                _closedByCode = true;
-                Close();
+                        _closedByCode = true;
+                        Close();
+                    }
+                    else
+                    {
+                        _telemetry.Write(module, "AutomaticUpdate", "Copy of 'Chem4WordUpdater.exe' to {userTempPath} failed.");
+                        MessageBox.Show($"Copy of 'Chem4WordUpdater.exe' to {userTempPath} failed. Please download {_downloadUrl} manually and install it.");
+                    }
+                }
+                else
+                {
+                    _telemetry.Write(module, "AutomaticUpdate", $"Folder '{userTempPath}' not found.");
+                    MessageBox.Show($"Folder '{userTempPath}' not found. Please download {_downloadUrl} manually and install it.");
+                }
             }
             catch (Exception ex)
             {
